+++ redirects.js
@@ -16,13 +16,12 @@ const setup = (app) => {
     .get((req, res) => res.redirect('/#narratives'));
 
   /* The URLs and nextstrain.org/narratives/ncov and nextstrain.org/narratives/ncov/sit-rep
-  aren't valid - redirect instead to the nCoV section of the splash page */
-  app.route(["/narratives/ncov", "/narratives/ncov/sit-rep"])
-    .get((req, res) => res.redirect('/ncov-sit-reps'));
+  aren't valid - redirect instead to the sit-reps section of the nCoV splash page */
+  app.route(["/narratives/ncov", "/narratives/ncov/sit-rep", "/ncov-sit-reps"])
+    .get((req, res) => res.redirect('/sars-cov-2#sit-reps'));
 
-  // TODO uncomment this reroute when we are ready
-  // app.route(["/ncov"])
-  // .get((req, res) => res.redirect('/sars-cov-2'));
+  app.route(["/ncov"])
+    .get((req, res) => res.redirect('/sars-cov-2'));
 
   /*
    * Redirect to translations of narratives if the client has
