+++ static-site/gatsby-node.js
@@ -193,16 +193,10 @@ exports.createPages = ({graphql, actions}) => {
           component: path.resolve("src/sections/users.jsx")
         });
 
-        // Create page detailing all SARS-CoV-2 builds
+        // Create page detailing all things SARS-CoV-2
         createPage({
           path: "/sars-cov-2",
           component: path.resolve("src/pages/sars-cov-2.jsx")
-        });
-
-        // Create page detailing all SARS-CoV-2 sit-reps
-        createPage({
-          path: "/ncov-sit-reps",
-          component: path.resolve("src/pages/ncov-sit-reps.jsx")
         });
 
         // search pages
